webhooks/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0

httpx==0.27.0

python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.3

google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.2

prometheus-client>=0.20.0
apscheduler>=3.10
sqlalchemy>=2.0.0
psycopg

pytest==7.4.3
pytest-asyncio==0.23.5